refactor: extract remaining routers and split test_api.py
